Use curl instead of wget for OSX portability

+++ download-data.sh
@@ -11,15 +11,16 @@ do
 	test -e "$cancer" && continue
 	mkdir "$cancer"
 	cd "$cancer"
-	wget "http://acgt.cs.tau.ac.il/multi_omic_benchmark/data/$cancer.zip"
+	curl -O "http://acgt.cs.tau.ac.il/multi_omic_benchmark/data/$cancer.zip"
 	unzip "$cancer.zip"
 	rm "$cancer.zip"
 	cd ..
 done
 
 cd ..
-test -e "clinical"  ||
-	wget http://acgt.cs.tau.ac.il/multi_omic_benchmark/data/clinical.zip &&
-	unzip clinical.zip &&
+if [ ! -e "clinical" ]
+then
+	curl -O http://acgt.cs.tau.ac.il/multi_omic_benchmark/data/clinical.zip
+	unzip clinical.zip
 	rm clinical.zip
-
+fi
